+++ message/index.js
@@ -15,32 +15,6 @@ const getStatus = require("./src/clients/getStatus");
 const getVersion = require("./src/clients/getVersion");
 const app = express();
 
-const { createLogger, format, transports } = require("winston");
-const logger = createLogger({
-  level: 'silly',
-  format: format.combine(
-    format.timestamp({
-      format: "YYYY-MM-DD HH:mm:ss"
-    }),
-    format.json()
-  ),
-  transports: [
-    new transports.Console({
-      level: "silly",
-      format: format.combine(
-        format.colorize(),
-        format.printf(
-          info => `${info.timestamp} ${info.level}: ${info.message}`
-        )
-      )
-    }),
-  ]
-});
-
-logger.info('Hello world');
-logger.warn('Warning message');
-logger.debug('Debugging info');
-
 const validator = new Validator({ allErrors: true });
 const { validate } = validator;
