+++ message/src/controllers/sendMessage.js
@@ -1,102 +1,110 @@
-require('dotenv').config()
+require("dotenv").config();
 const http = require("http");
 const saveMessage = require("../clients/saveMessage");
 const rollBackQueue = require("./rollBackQueue");
-const braker = require('../circuitBreaker');
+const braker = require("../circuitBreaker");
 const util = require("util");
+const logger = require("../winston");
 
-module.exports = function (message, done) {
-    console.log(message)
-    const body = JSON.stringify(message);
-    const idQuery = message.qId
+module.exports = function(message, done) {
+  console.log(message);
+  const body = JSON.stringify(message);
+  const idQuery = message.qId;
 
-    if (message.payment === true) {
-      const postOptions = {
-        host: process.env.MESSAGEAPP_HOST_LOCAL,
-        port: process.env.MESSAGEAPP_PORT,
-        path: "/message",
-        method: "post",
-        json: true,
-        headers: {
-          "Content-Type": "application/json",
-          "Content-Length": Buffer.byteLength(body)
-        }
-      };
+  if (message.payment === true) {
+    const postOptions = {
+      host: process.env.MESSAGEAPP_HOST_LOCAL,
+      port: process.env.MESSAGEAPP_PORT,
+      path: "/message",
+      method: "post",
+      json: true,
+      headers: {
+        "Content-Type": "application/json",
+        "Content-Length": Buffer.byteLength(body)
+      }
+    };
 
-      function messageappCircuitBraker(postOptions) {
-        return new Promise(function (resolve, reject) {
+    function messageappCircuitBraker(postOptions) {
+      return new Promise(function(resolve, reject) {
+        let postReq = http.request(postOptions);
 
-          let postReq = http.request(postOptions);
-
-          postReq.on("response", postRes => {
-            if (postRes.statusCode === 200) {
-              saveMessage({
-                  ...message,
-                  qId: message.qId,
-                  status: "OK"
-                },
-                () => {
-                  console.log("Error in server response");
-                }, idQuery
-              )
-              resolve("OK")
-            } else {
-              console.error("Error while sending message");
-
-              saveMessage({
-                  ...message,
-                  qId: message.qId,
-                  status: "ERROR",
-                  payment: false
-                },
-                () => {
-                  rollBackQueue()
-                  console.log("Internal server error: SERVICE ERROR");
-                }, idQuery
-              )
-              reject("ERROR")
-            }
-          });
-          postReq.setTimeout(1000);
-          postReq.on("timeout", () => {
-            console.error("Timeout Exceeded!");
-            postReq.abort();
-            saveMessage({
+        postReq.on("response", postRes => {
+          if (postRes.statusCode === 200) {
+            saveMessage(
+              {
                 ...message,
                 qId: message.qId,
-                status: "TIMEOUT"
+                status: "OK"
               },
               () => {
-                console.log("Internal server error: TIMEOUT");
-              }, idQuery
-            )
-            reject("TIMEOUT")
-          });
-          postReq.on("error", () => {});
+                logger.info("Acepted");
+              },
+              idQuery
+            );
+            resolve("OK");
+          } else {
+            logger.error("Error while sending message");
 
-          postReq.write(body);
-          postReq.end();
-        })
-      }
-      const circuit = braker.slaveCircuit(messageappCircuitBraker)
-      circuit
-        .exec(postOptions)
-        .then(result => {
-          console.log(`result: ${result}`);
-        })
-        .catch(err => {
-          console.error(`${err}`);
+            saveMessage(
+              {
+                ...message,
+                qId: message.qId,
+                status: "ERROR",
+                payment: false
+              },
+              () => {
+                rollBackQueue();
+                logger.warn("Internal server error: SERVICE ERROR");
+              },
+              idQuery
+            );
+            reject("ERROR");
+          }
         });
-        done()
-    } else {
-      saveMessage({
-          ...message,
-          qId: message.qId,
-          status: "INSUFICIENT CREDIT"
-        },
-        () => {
-          console.log("Internal server error:INSUFICIENT CREDIT");
-        }, idQuery
-      )
+        postReq.setTimeout(1000);
+        postReq.on("timeout", () => {
+          logger.info("Timeout Exceeded!");
+          postReq.abort();
+          saveMessage(
+            {
+              ...message,
+              qId: message.qId,
+              status: "TIMEOUT"
+            },
+            () => {
+              logger.warn("Internal server error: TIMEOUT");
+            },
+            idQuery
+          );
+          reject("TIMEOUT");
+        });
+        postReq.on("error", () => {});
+
+        postReq.write(body);
+        postReq.end();
+      });
     }
+    const circuit = braker.slaveCircuit(messageappCircuitBraker);
+    circuit
+      .exec(postOptions)
+      .then(result => {
+        logger.info(`result: ${result}`);
+      })
+      .catch(err => {
+        logger.error(`${err}`);
+      });
+    done();
+  } else {
+    saveMessage(
+      {
+        ...message,
+        qId: message.qId,
+        status: "INSUFICIENT CREDIT"
+      },
+      () => {
+        logger.info("Internal server error:INSUFICIENT CREDIT");
+      },
+      idQuery
+    );
   }
+};
